refactor: put SettingsModalContent (with error boundary) and SettingModal in the same flie
